Add ros integration to makefile
A roscore option has been added to the makefile and the user can now
pass in the name of the instance of the docker image to allow multiple
versions to be created. All these instances are now connected by a ROS
network allowing them to publish/subscribe between each other.